webapp2_std_google_cloud_storage/requirements.txt
webapp2>=3.0.0b1
google-cloud-storage
